feat: new project, learn the bisection by finding the square root of a number
Numerical methods are used to approximate solutions to mathematical problems that are difficult or impossible to solve analytically.

In this project, you will explore the numerical method of bisection to find the square root of a number by iteratively narrowing down the possible range of values that contain the square root.